Renamed a few packages to internal in the org.eclipse.smarthome.io.rest.core bundle. See openhab/static-code-analysis#100
