+++ bundles/io/org.eclipse.smarthome.io.rest.core/OSGI-INF/bindingresource.xml
@@ -8,7 +8,7 @@
     http://www.eclipse.org/legal/epl-v10.html
 
 -->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.smarthome.io.rest.core.binding.BindingResource">
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.smarthome.io.rest.core.internal.binding.BindingResource">
    <implementation class="org.eclipse.smarthome.io.rest.core.binding.BindingResource"/>
    <service>
       <provide interface="org.eclipse.smarthome.io.rest.core.binding.BindingResource"/>
