+++ bundles/io/org.eclipse.smarthome.io.rest.core/OSGI-INF/channeltyperesource.xml
@@ -11,7 +11,7 @@
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.smarthome.io.rest.core.thing.ChannelTypeResource">
    <implementation class="org.eclipse.smarthome.io.rest.core.channel.ChannelTypeResource"/>
    <service>
-      <provide interface="org.eclipse.smarthome.io.rest.core.channel.ChannelTypeResource"/>
+      <provide interface="org.eclipse.smarthome.io.rest.core.internal.channel.ChannelTypeResource"/>
       <provide interface="org.eclipse.smarthome.io.rest.RESTResource"/>
    </service>
    <reference bind="setConfigDescriptionRegistry" cardinality="0..1" interface="org.eclipse.smarthome.config.core.ConfigDescriptionRegistry" name="ConfigDescriptionRegistry" policy="dynamic" unbind="unsetConfigDescriptionRegistry"/>
