+++ bundles/io/org.eclipse.smarthome.io.rest.core/OSGI-INF/configurableserviceresource.xml
@@ -9,10 +9,10 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.smarthome.io.rest.core.service.ConfigurableServiceResource">
-   <implementation class="org.eclipse.smarthome.io.rest.core.service.ConfigurableServiceResource"/>
+   <implementation class="org.eclipse.smarthome.io.rest.core.internal.service.ConfigurableServiceResource"/>
    <service>
       <provide interface="org.eclipse.smarthome.io.rest.RESTResource"/>
-      <provide interface="org.eclipse.smarthome.io.rest.core.service.ConfigurableServiceResource"/>
+      <provide interface="org.eclipse.smarthome.io.rest.core.internal.service.ConfigurableServiceResource"/>
    </service>
    <reference bind="setConfigurationService" cardinality="0..1" interface="org.eclipse.smarthome.io.rest.core.config.ConfigurationService" name="ConfigurationService" policy="dynamic" unbind="unsetConfigurationService"/>
    <reference bind="setConfigDescriptionRegistry" cardinality="0..1" interface="org.eclipse.smarthome.config.core.ConfigDescriptionRegistry" name="ConfigDescriptionRegistry" policy="dynamic" unbind="unsetConfigDescriptionRegistry"/>
