+++ bundles/io/org.eclipse.smarthome.io.rest.core/OSGI-INF/discoveryresource.xml
@@ -8,10 +8,10 @@
     http://www.eclipse.org/legal/epl-v10.html
 
 -->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.smarthome.io.rest.core.discovery.DiscoveryResource">
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.smarthome.io.rest.core.internal.discovery.DiscoveryResource">
    <implementation class="org.eclipse.smarthome.io.rest.core.discovery.DiscoveryResource"/>
    <service>
-      <provide interface="org.eclipse.smarthome.io.rest.core.discovery.DiscoveryResource"/>
+      <provide interface="org.eclipse.smarthome.io.rest.core.internal.discovery.DiscoveryResource"/>
       <provide interface="org.eclipse.smarthome.io.rest.RESTResource"/>
    </service>
    <reference bind="setDiscoveryServiceRegistry" cardinality="0..1" interface="org.eclipse.smarthome.config.discovery.DiscoveryServiceRegistry" name="DiscoveryServiceRegistry" policy="dynamic" unbind="unsetDiscoveryServiceRegistry"/>
