+++ bundles/io/org.eclipse.smarthome.io.rest.core/OSGI-INF/itemchannellinkresource.xml
@@ -9,10 +9,10 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.smarthome.io.rest.core.link.ItemChannelLinkResource">
-   <implementation class="org.eclipse.smarthome.io.rest.core.link.ItemChannelLinkResource"/>
+   <implementation class="org.eclipse.smarthome.io.rest.core.internal.link.ItemChannelLinkResource"/>
    <service>
       <provide interface="org.eclipse.smarthome.io.rest.RESTResource"/>
-      <provide interface="org.eclipse.smarthome.io.rest.core.link.ItemChannelLinkResource"/>
+      <provide interface="org.eclipse.smarthome.io.rest.core.internal.link.ItemChannelLinkResource"/>
    </service>
    <reference bind="setItemChannelLinkRegistry" cardinality="0..1" interface="org.eclipse.smarthome.core.thing.link.ItemChannelLinkRegistry" name="ItemChannelLinkRegistry" policy="dynamic" unbind="unsetItemChannelLinkRegistry"/>
    <reference bind="setThingLinkManager" cardinality="0..1" interface="org.eclipse.smarthome.core.thing.link.ThingLinkManager" name="ThingLinkManager" policy="dynamic" unbind="unsetThingLinkManager"/>
