+++ bundles/io/org.eclipse.smarthome.io.rest.core/OSGI-INF/itemresource.xml
@@ -11,7 +11,7 @@
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.io.rest.core.item">
    <implementation class="org.eclipse.smarthome.io.rest.core.item.ItemResource"/>
    <service>
-      <provide interface="org.eclipse.smarthome.io.rest.core.item.ItemResource"/>
+      <provide interface="org.eclipse.smarthome.io.rest.core.internal.item.ItemResource"/>
       <provide interface="org.eclipse.smarthome.io.rest.RESTResource"/>
    </service>
    <reference bind="setItemRegistry" cardinality="0..1" interface="org.eclipse.smarthome.core.items.ItemRegistry" name="ItemRegistry" policy="dynamic" unbind="unsetItemRegistry"/>
