+++ bundles/io/org.eclipse.smarthome.io.rest.core/OSGI-INF/thingtyperesource.xml
@@ -9,9 +9,9 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.smarthome.io.rest.core.thing.ThingTypeResource">
-   <implementation class="org.eclipse.smarthome.io.rest.core.thing.ThingTypeResource"/>
+   <implementation class="org.eclipse.smarthome.io.rest.core.internal.thing.ThingTypeResource"/>
    <service>
-      <provide interface="org.eclipse.smarthome.io.rest.core.thing.ThingTypeResource"/>
+      <provide interface="org.eclipse.smarthome.io.rest.core.internal.thing.ThingTypeResource"/>
       <provide interface="org.eclipse.smarthome.io.rest.RESTResource"/>
    </service>
    <reference bind="setConfigDescriptionRegistry" cardinality="0..1" interface="org.eclipse.smarthome.config.core.ConfigDescriptionRegistry" name="ConfigDescriptionRegistry" policy="dynamic" unbind="unsetConfigDescriptionRegistry"/>
