+++ bundles/io/org.eclipse.smarthome.io.rest.core/src/main/java/org/eclipse/smarthome/io/rest/core/internal/discovery/DiscoveryResource.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.io.rest.core.discovery;
+package org.eclipse.smarthome.io.rest.core.internal.discovery;
 
 import java.util.Collection;
 import java.util.LinkedHashSet;
